+++ src/parse.rs
@@ -596,6 +596,8 @@ impl<'a> Parser<'a> {
             (end, Expr::ContinueFromPreviousMatchEnd)
         } else if b == b'O' && !in_class {
             (end, Expr::Any { newline: true })
+        } else if b == b'N' && !in_class {
+            (end, Expr::Any { newline: false })
         } else if b == b'g' && !in_class {
             if end == self.re.len() {
                 return Err(Error::ParseError(